multiblockchain: Better abstract over `TriggerData`, `MappingTrigger` and `DataSource`
