+++ chain/ethereum/src/lib.rs
@@ -8,11 +8,14 @@ mod transport;
 pub use self::ethereum_adapter::EthereumAdapter;
 pub use self::transport::{EventLoopHandle, Transport};
 
-// ETHDEP: This concrete type should probably not be exposed.
+// ETHDEP: These concrete types should probably not be exposed.
 pub use data_source::DataSource;
+pub use trigger::MappingTrigger;
 
 pub mod chain;
+
 mod network;
+mod trigger;
 
 pub use crate::adapter::{
     EthereumAdapter as EthereumAdapterTrait, EthereumContractCall, EthereumContractCallError,
@@ -20,3 +23,6 @@ pub use crate::adapter::{
 };
 pub use crate::chain::{Chain, WrappedBlockFinality};
 pub use crate::network::EthereumNetworks;
+
+#[cfg(test)]
+mod tests;
