+++ graph/src/components/ethereum/types.rs
@@ -1,9 +1,6 @@
 use ethabi::LogParam;
 use serde::{Deserialize, Serialize};
-use slog::{o, SendSyncRefUnwindSafeKV};
-use std::sync::Arc;
-use std::{cmp::Ordering, convert::TryFrom};
-use strum_macros::AsStaticStr;
+use std::convert::TryFrom;
 use web3::types::{
     Action, Address, Block, Bytes, Log, Res, Trace, Transaction, TransactionReceipt, H160, H256,
     U128, U256, U64,
@@ -11,11 +8,7 @@ use web3::types::{
 
 use crate::{
     blockchain::BlockPtr,
-    cheap_clone::CheapClone,
-    prelude::{
-        BlockNumber, DeploymentHash, EntityKey, MappingBlockHandler, MappingCallHandler,
-        MappingEventHandler, ToEntityKey,
-    },
+    prelude::{BlockNumber, DeploymentHash, EntityKey, ToEntityKey},
 };
 
 pub type LightEthereumBlock = Block<Transaction>;
@@ -173,209 +166,6 @@ impl EthereumCall {
             transaction_hash: trace.transaction_hash,
             transaction_index,
         })
-    }
-}
-
-#[derive(Clone, Debug)]
-pub enum EthereumTrigger {
-    Block(BlockPtr, EthereumBlockTriggerType),
-    Call(Arc<EthereumCall>),
-    Log(Arc<Log>),
-}
-
-impl PartialEq for EthereumTrigger {
-    fn eq(&self, other: &Self) -> bool {
-        match (self, other) {
-            (Self::Block(a_ptr, a_kind), Self::Block(b_ptr, b_kind)) => {
-                a_ptr == b_ptr && a_kind == b_kind
-            }
-
-            (Self::Call(a), Self::Call(b)) => a == b,
-
-            (Self::Log(a), Self::Log(b)) => {
-                a.transaction_hash == b.transaction_hash && a.log_index == b.log_index
-            }
-
-            _ => false,
-        }
-    }
-}
-
-impl Eq for EthereumTrigger {}
-
-#[derive(Clone, Debug, PartialEq, Eq)]
-pub enum EthereumBlockTriggerType {
-    Every,
-    WithCallTo(Address),
-}
-
-impl EthereumTrigger {
-    pub fn block_number(&self) -> BlockNumber {
-        match self {
-            EthereumTrigger::Block(block_ptr, _) => block_ptr.number,
-            EthereumTrigger::Call(call) => call.block_number,
-            EthereumTrigger::Log(log) => i32::try_from(log.block_number.unwrap().as_u64()).unwrap(),
-        }
-    }
-
-    pub fn block_hash(&self) -> H256 {
-        match self {
-            EthereumTrigger::Block(block_ptr, _) => block_ptr.hash_as_h256(),
-            EthereumTrigger::Call(call) => call.block_hash,
-            EthereumTrigger::Log(log) => log.block_hash.unwrap(),
-        }
-    }
-}
-
-impl Ord for EthereumTrigger {
-    fn cmp(&self, other: &Self) -> Ordering {
-        match (self, other) {
-            // Keep the order when comparing two block triggers
-            (Self::Block(..), Self::Block(..)) => Ordering::Equal,
-
-            // Block triggers always come last
-            (Self::Block(..), _) => Ordering::Greater,
-            (_, Self::Block(..)) => Ordering::Less,
-
-            // Calls are ordered by their tx indexes
-            (Self::Call(a), Self::Call(b)) => a.transaction_index.cmp(&b.transaction_index),
-
-            // Events are ordered by their log index
-            (Self::Log(a), Self::Log(b)) => a.log_index.cmp(&b.log_index),
-
-            // Calls vs. events are logged by their tx index;
-            // if they are from the same transaction, events come first
-            (Self::Call(a), Self::Log(b))
-                if a.transaction_index == b.transaction_index.unwrap().as_u64() =>
-            {
-                Ordering::Greater
-            }
-            (Self::Log(a), Self::Call(b))
-                if a.transaction_index.unwrap().as_u64() == b.transaction_index =>
-            {
-                Ordering::Less
-            }
-            (Self::Call(a), Self::Log(b)) => a
-                .transaction_index
-                .cmp(&b.transaction_index.unwrap().as_u64()),
-            (Self::Log(a), Self::Call(b)) => a
-                .transaction_index
-                .unwrap()
-                .as_u64()
-                .cmp(&b.transaction_index),
-        }
-    }
-}
-
-impl PartialOrd for EthereumTrigger {
-    fn partial_cmp(&self, other: &Self) -> Option<Ordering> {
-        Some(self.cmp(other))
-    }
-}
-
-#[derive(AsStaticStr)]
-pub enum MappingTrigger {
-    Log {
-        block: Arc<LightEthereumBlock>,
-        transaction: Arc<Transaction>,
-        log: Arc<Log>,
-        params: Vec<LogParam>,
-        handler: MappingEventHandler,
-    },
-    Call {
-        block: Arc<LightEthereumBlock>,
-        transaction: Arc<Transaction>,
-        call: Arc<EthereumCall>,
-        inputs: Vec<LogParam>,
-        outputs: Vec<LogParam>,
-        handler: MappingCallHandler,
-    },
-    Block {
-        block: Arc<LightEthereumBlock>,
-        handler: MappingBlockHandler,
-    },
-}
-
-// Logging the block is too verbose, so this strips the block from the trigger for Debug.
-impl std::fmt::Debug for MappingTrigger {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        #[derive(Debug)]
-        pub enum MappingTriggerWithoutBlock {
-            Log {
-                transaction: Arc<Transaction>,
-                log: Arc<Log>,
-                params: Vec<LogParam>,
-                handler: MappingEventHandler,
-            },
-            Call {
-                transaction: Arc<Transaction>,
-                call: Arc<EthereumCall>,
-                inputs: Vec<LogParam>,
-                outputs: Vec<LogParam>,
-                handler: MappingCallHandler,
-            },
-            Block {
-                handler: MappingBlockHandler,
-            },
-        }
-
-        let trigger_without_block = match self {
-            MappingTrigger::Log {
-                block: _,
-                transaction,
-                log,
-                params,
-                handler,
-            } => MappingTriggerWithoutBlock::Log {
-                transaction: transaction.cheap_clone(),
-                log: log.cheap_clone(),
-                params: params.clone(),
-                handler: handler.clone(),
-            },
-            MappingTrigger::Call {
-                block: _,
-                transaction,
-                call,
-                inputs,
-                outputs,
-                handler,
-            } => MappingTriggerWithoutBlock::Call {
-                transaction: transaction.cheap_clone(),
-                call: call.cheap_clone(),
-                inputs: inputs.clone(),
-                outputs: outputs.clone(),
-                handler: handler.clone(),
-            },
-            MappingTrigger::Block { block: _, handler } => MappingTriggerWithoutBlock::Block {
-                handler: handler.clone(),
-            },
-        };
-
-        write!(f, "{:?}", trigger_without_block)
-    }
-}
-
-impl MappingTrigger {
-    pub fn handler_name(&self) -> &str {
-        match self {
-            MappingTrigger::Log { handler, .. } => &handler.handler,
-            MappingTrigger::Call { handler, .. } => &handler.handler,
-            MappingTrigger::Block { handler, .. } => &handler.handler,
-        }
-    }
-
-    pub fn logging_extras(&self) -> impl SendSyncRefUnwindSafeKV {
-        match self {
-            MappingTrigger::Log { handler, log, .. } => o! {
-                "signature" => handler.event.to_string(),
-                "address" => format!("{}", &log.address),
-            },
-            MappingTrigger::Call { handler, call, .. } => o! {
-                "function" => handler.function.to_string(),
-                "to" => format!("{}", &call.to),
-            },
-            MappingTrigger::Block { .. } => o! { "" => String::new(), "" => String::new() },
-        }
     }
 }
