+++ runtime/wasm/src/host.rs
@@ -5,20 +5,18 @@ use std::time::{Duration, Instant};
 use async_trait::async_trait;
 use futures::sync::mpsc::Sender;
 use futures03::channel::oneshot::channel;
-use graph::{
-    blockchain::{Blockchain, DataSource},
-    components::store::CallCache,
-};
-use strum::AsStaticRef as _;
-
 use graph::components::arweave::ArweaveAdapter;
-use graph::components::ethereum::*;
 use graph::components::store::SubgraphStore;
 use graph::components::subgraph::{MappingError, SharedProofOfIndexing};
 use graph::components::three_box::ThreeBoxAdapter;
 use graph::prelude::{
     RuntimeHost as RuntimeHostTrait, RuntimeHostBuilder as RuntimeHostBuilderTrait, *,
 };
+use graph::{
+    blockchain::{Blockchain, DataSource, MappingTrigger as _},
+    components::store::CallCache,
+};
+use graph_chain_ethereum::MappingTrigger;
 use graph_chain_ethereum::{EthereumAdapterTrait, EthereumNetworks};
 
 use crate::mapping::{MappingContext, MappingRequest};
@@ -93,7 +91,7 @@ where
     CC: CallCache,
     C: Blockchain<
         Block = graph_chain_ethereum::WrappedBlockFinality,
-        TriggerData = EthereumTrigger,
+        MappingTrigger = graph_chain_ethereum::MappingTrigger,
         DataSource = graph_chain_ethereum::DataSource,
     >,
 {
@@ -174,7 +172,7 @@ pub struct RuntimeHost<C: Blockchain> {
 
 impl<C> RuntimeHost<C>
 where
-    C: Blockchain<DataSource = graph_chain_ethereum::DataSource>,
+    C: Blockchain,
 {
     fn new(
         ethereum_adapter: Arc<dyn EthereumAdapterTrait>,
@@ -223,14 +221,12 @@ where
         block_ptr: BlockPtr,
         proof_of_indexing: SharedProofOfIndexing,
     ) -> Result<BlockState, MappingError> {
-        let trigger_type = trigger.as_static();
         let handler = trigger.handler_name().to_string();
 
         let extras = trigger.logging_extras();
         trace!(
             logger, "Start processing Ethereum trigger";
             &extras,
-            "trigger_type" => trigger_type,
             "handler" => &handler,
             "data_source" => &self.data_source.name(),
         );
@@ -266,7 +262,6 @@ where
         info!(
             logger, "Done processing Ethereum trigger";
             &extras,
-            "trigger_type" => trigger_type,
             "total_ms" => elapsed.as_millis(),
             "handler" => handler,
             "data_source" => &self.data_source.name(),
@@ -279,18 +274,14 @@ where
 #[async_trait]
 impl<C> RuntimeHostTrait<C> for RuntimeHost<C>
 where
-    C: Blockchain<
-        Block = graph_chain_ethereum::WrappedBlockFinality,
-        TriggerData = EthereumTrigger,
-        DataSource = graph_chain_ethereum::DataSource,
-    >,
+    C: Blockchain<MappingTrigger = graph_chain_ethereum::MappingTrigger>,
 {
     fn match_and_decode(
         &self,
-        trigger: &EthereumTrigger,
-        block: Arc<LightEthereumBlock>,
+        trigger: &C::TriggerData,
+        block: Arc<C::Block>,
         logger: &Logger,
-    ) -> Result<Option<MappingTrigger>, Error> {
+    ) -> Result<Option<C::MappingTrigger>, Error> {
         self.data_source.match_and_decode(trigger, block, logger)
     }
 
@@ -298,7 +289,7 @@ where
         &self,
         logger: &Logger,
         block_ptr: BlockPtr,
-        trigger: MappingTrigger,
+        trigger: C::MappingTrigger,
         state: BlockState,
         proof_of_indexing: SharedProofOfIndexing,
     ) -> Result<BlockState, MappingError> {
